Split server test file into multiple test files
