+++ services/server/package.json
@@ -14,9 +14,10 @@
     "postgres-test:start": "docker-compose -f test/docker-compose.yml up -d",
     "postgres-test:stop": "docker-compose -f test/docker-compose.yml rm -fsv",
     "postgres-test:migrate": "cd ../database && npm run migrate:up -- --env test",
-    "test": "npm run postgres-test:migrate && NODE_CONFIG_DIR=./dist/config DOTENV_CONFIG_PATH=./.env c8 --reporter=none mocha -r dotenv/config --exit",
-    "test-local": "export DOCKER_HOST_POSTGRES_TEST_PORT=${DOCKER_HOST_POSTGRES_TEST_PORT:-5431} && npm run postgres-test:start && echo \"Sleeping 5 seconds to wait for the container to start\" && sleep 5 && npm run postgres-test:migrate && NODE_CONFIG_DIR=./dist/config DOTENV_CONFIG_PATH=./.env c8 --reporter=none mocha -r dotenv/config --exit; status=$?; npm run postgres-test:stop; exit $status",
-    "test-local:repository": "export DOCKER_HOST_POSTGRES_TEST_PORT=${DOCKER_HOST_POSTGRES_TEST_PORT:-5431} && npm run postgres-test:start && echo \"Sleeping 5 seconds to wait for the container to start\" && sleep 5 && npm run postgres-test:migrate && NODE_CONFIG_DIR=./dist/config DOTENV_CONFIG_PATH=./.env.test c8 --reporter=none mocha -r dotenv/config --exit test/integration/handlers/repository.spec.ts; status=$?; npm run postgres-test:stop; exit $status",
+    "test:unit": "NODE_CONFIG_DIR=./src/config DOTENV_CONFIG_PATH=./.env.test c8 --reporter=none mocha -r dotenv/config --exit test/unit/**",
+    "test:integration": "npm run postgres-test:migrate && NODE_CONFIG_DIR=./src/config DOTENV_CONFIG_PATH=./.env.test c8 --reporter=none mocha -r dotenv/config --exit test/integration/**",
+    "test-local:integration": "export DOCKER_HOST_POSTGRES_TEST_PORT=${DOCKER_HOST_POSTGRES_TEST_PORT:-5431} && npm run postgres-test:start && echo \"Sleeping 5 seconds to wait for the container to start\" && sleep 5 && npm run test:integration; status=$?; npm run postgres-test:stop; exit $status",
+    "test-local": "npm run test:unit && npm run test-local:integration",
     "test:chains": "mocha test/chains/chain-tests.spec.ts --reporter mochawesome --reporter-options reportDir=chain-tests-report,reportFilename=report --exit",
     "lint": "eslint . --ext .ts",
     "coverage": "node ./node_modules/c8/bin/c8.js --reporter=lcov --reporter=text-summary ./node_modules/tape/bin/tape ./test/index.js",
